requirements.txt
# Core dependencies
python-telegram-bot[all]==20.7
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
aiosqlite==0.19.0
python-dotenv==1.0.0

# Date and time handling
pytz==2023.3
python-dateutil==2.8.2

# Data validation
pydantic==2.5.0

# Utilities
aiofiles==23.2.1
phonenumbers==8.13.26

# Async utilities
asyncio==3.4.3
apscheduler==3.10.4

# For PDF/Image handling
Pillow==10.1.0

# Logging
structlog==23.2.0

# Google Sheets integration
gspread==5.12.0
oauth2client==4.1.3
